+++ finagle-redis/src/test/scala/com/twitter/finagle/redis/integration/ClientServerIntegrationSpec.scala
@@ -387,7 +387,7 @@ class ClientServerIntegrationSpec extends SpecificationWithJUnit {
         val key = "setex"
         client(SetEx(key, 10, "Hello"))() mustEqual StatusReply("OK")
         client(Ttl(key))() match {
-          case IntegerReply(seconds) => seconds must beCloseTo(10, 2)
+          case IntegerReply(seconds) => seconds.toInt must beCloseTo(10, 2)
           case _ => fail("Expected IntegerReply")
         }
         assertBulkReply(client(Get(key)), "Hello")
@@ -439,7 +439,7 @@ class ClientServerIntegrationSpec extends SpecificationWithJUnit {
   }
 
   def assertIntegerReply(reply: Future[Reply], expects: Int, delta: Int = 10) = reply() match {
-    case IntegerReply(amnt) => amnt must beCloseTo(expects, delta)
+    case IntegerReply(amnt) => amnt.toInt must beCloseTo(expects, delta)
     case _ => fail("Expected IntegerReply")
   }
